Create Find Merge Point of Two Lists
The problem can be found here: https://www.hackerrank.com/challenges/find-the-merge-point-of-two-joined-linked-lists/problem 
Note: this has failed the 2nd test case:

5
0
1
9
72
2
6
6
2
5
3
2
2
9
3
10
4
1
5
9
9
5
1
8
1
6
6
9
6
8
10
10
8
3
8
1
7
6
7
8
10
9
9
1
9
3
3
4
3
7
4
1
3
4
1
5
1
8
5
6
5
6
7
5
1
5
5
7
2
3
4
3
3
5
5
1
7
71
77
10
5
8
2
9
2
2
5
2
10
1
10
6
7
6
10
7
2
7
5
3
9
8
5
2
2
10
2
1
8
8
2
5
6
3
3
9
5
10
2
6
10
1
1
8
7
3
6
8
9
10
1
9
8
7
2
1
8
6
3
6
3
4
2
10
7
4
8
3
5
10
10
5
10
2
2
6
94
8
6
4
7
6
5
6
4
6
9
2
1
1
9
4
5
2
3
3
5
3
7
2
2
6
6
1
7
7
9
3
6
4
6
3
1
10
10
6
8
8
9
8
1
9
3
5
10
6
9
7
8
5
8
1
10
3
3
8
1
1
10
7
6
8
1
8
9
10
4
6
8
2
4
8
1
8
4
10
3
4
8
2
8
5
4
9
7
7
6
10
9
8
6
6
35
6
2
5
5
7
3
4
9
8
3
1
5
8
2
10
1
10
3
10
4
7
8
3
3
6
2
3
5
9
7
1
4
1
7
10
89
1
6
7
8
10
9
3
8
1
4
10
10
6
10
5
2
9
7
6
6
10
1
10
8
9
2
3
9
1
5
10
1
10
6
9
1
7
3
8
7
6
10
8
3
1
2
7
9
1
2
7
10
4
6
10
3
10
2
3
10
6
2
10
7
10
10
8
6
2
7
4
9
6
1
2
8
4
8
7
6
1
3
8
5
10
7
7
1
10
88
89
6
3
2
4
2
1
1
7
5
8
10
3
5
2
6
5
7
5
3
3
6
7
10
2
6
8
10
7
7
8
8
4
10
9
8
3
2
10
9
6
7
10
10
4
1
8
8
10
2
10
2
9
8
3
10
3
10
9
1
8
9
9
2
10
9
1
3
10
10
3
7
9
3
9
4
5
6
1
4
9
2
7
8
9
9
7
3
10
8
66
8
8
6
1
7
4
1
1
6
2
4
2
10
8
2
3
4
7
5
10
6
8
6
5
8
7
3
3
6
10
8
5
7
3
7
6
8
9
6
3
1
1
7
2
10
8
5
4
7
1
5
4
9
10
8
6
8
10
8
6
2
7
10
8
9
9
10
43
7
1
1
9
3
9
1
3
7
5
8
5
5
2
8
5
3
7
3
1
6
2
6
7
9
7
7
9
5
1
7
4
3
8
2
6
6
2
8
4
8
7
10
55
8
7
9
2
5
1
4
1
3
1
9
3
10
7
1
6
8
8
9
2
7
3
7
4
6
6
10
4
2
9
8
1
8
8
5
2
9
8
2
3
1
3
5
10
1
5
5
8
4
6
2
10
8
8
6

Given pointers to the head nodes of  linked lists that merge together at some point, find the Node where the two lists merge. It is guaranteed that the two head Nodes will be different, and neither will be NULL.

In the diagram below, the two lists converge at Node x:

[List #1] a--->b--->c
                     \
                      x--->y--->z--->NULL
                     /
     [List #2] p--->q
Complete the int findMergeNode(SinglyLinkedListNode* head1, SinglyLinkedListNode* head2) method so that it finds and returns the data value of the Node where the two lists merge.

Input Format

Do not read any input from stdin/console.

The findMergeNode(SinglyLinkedListNode,SinglyLinkedListNode) method has two parameters,  and , which are the non-null head Nodes of two separate linked lists that are guaranteed to converge.

Constraints

The lists will merge. 
. 
 .

Output Format

Do not write any output to stdout/console.

Each Node has a data field containing an integer. Return the integer data for the Node where the two lists merge.

Sample Input

The diagrams below are graphical representations of the lists that input Nodes  and  are connected to. Recall that this is a method-only challenge; the method only has initial visibility to those  Nodes and must explore the rest of the Nodes using some algorithm of your own design.

Test Case 0

 1
  \
   2--->3--->NULL
  /
 1
Test Case 1

1--->2
      \
       3--->Null
      /
     1
Sample Output

2
3
Explanation

Test Case 0: As demonstrated in the diagram above, the merge Node's data field contains the integer . 
Test Case 1: As demonstrated in the diagram above, the merge Node's data field contains the integer .