Update Tree: Preorder Traversal
Complete the preOrder function in your editor below, which has  1 parameter: a pointer to the root of a binary tree. It must print the values in the tree's preorder traversal as a single line of space-separated values.

Input Format: A hidden tester code passes the root node of a binary tree to your preOrder function.

Constraints:  1 <= Nodes in the tree <= 100

Output: Print the tree's preorder traversal as a single line of space-separated values.