Create Tree: Postorder Traversal
Problem:
Complete the postOrder function in your editor below, which has  1 parameter: a pointer to the root of a binary tree. It must print the values in the tree's postorder traversal as a single line of space-separated values.

Input Format: 1 <= Nodes in the tree  <= 500
Output Format: Print the tree's postorder traversal as a single line of space-separated values.