Tree: Inorder Traversal
Problem:
Complete the inOrder function in your editor below, which has  1 parameter: a pointer to the root of a binary tree. It must print the values in the tree's inorder traversal as a single line of space-separated values.

Input Format: Our hidden tester code passes the root node of a binary tree to your inOrder function.

Constraints: 1 <= Nodes in the tree  <= 500

Output Format: Print the tree's inorder traversal as a single line of space-separated values.